fix(nvim): error when closing gitcommit buffer too fast
It was due to languagetool gets triggered right after switching from
INSERT to NORMAL mode. If we close the buffer before languagetool
finishes, it won't be able to find the buffer and will report an error
on that.

I tried but couldn't find any elegant way to ignore the error. The error
is not important to me, so I apply the fix here, to get rid of the
message.